Gesture Anywhere: Add gesture anywhere feature into the ChaOS Lab [1/2]
When enabled a user can activate the gesture anywhere view by touching
the designated trigger area.  If a gesture is recognized as one the user
has created and assigned to an action, that action will be performed.
Since this is part of the system UI and implemented as an overlay, it
will be available and usable system wide, with the exception of the
keyguard for obvious reasons.

Change-Id: Icc2822e1de276bfb2850e885a84edab1ca84cf0d

Gesture Anywhere: Adjust position on orientation change.

Also added hiding home, recent and search from navbar when
showing the GestureAnywhere view.

Change-Id: I1dd13f38af265d6966441f2cd99b78ea9b3d4d90

Gesture anywhere: Collapse if expanded when screen turns off.

Change-Id: I85c749182dab9e478d7337b3556c362eaa7dd74d

ChaosLab: Add annotation to be used when adding new features to the lab.

This should be used when adding completely new features to ChameleonOS,
and not for simple things that expand upon a current feature.

Change-Id: I37d7250424967d41ecfb885120627f5397b2f804

Gesture Anywhere: Adjust position on orientation change.

Also added hiding home, recent and search from navbar when
showing the GestureAnywhere view.

Change-Id: I1dd13f38af265d6966441f2cd99b78ea9b3d4d90

Conflicts:
	core/java/android/provider/Settings.java

Conflicts:
	core/java/android/provider/Settings.java